# Timbangan/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Timbangan Digital IoT API",
        "description": "API untuk integrasi perangkat timbangan IoT (Arduino/ESP8266) dengan sistem penimbangan digital. Semua endpoint menggunakan token perangkat untuk autentikasi.",
        "contact": {
            "name": "Timbangan Digital Team"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": " Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/fg-pasuruan/settings": {
            "get": {
                "tags": [
                    "FG Pasuruan"
                ],
                "summary": "Get pengaturan sesi aktif timbangan FG Pasuruan",
                "description": "Dipanggil oleh Arduino/ESP8266 untuk mendapatkan informasi sesi aktif (produk, operator, kode produksi). Gunakan token perangkat sebagai query parameter.",
                "operationId": "67b2040c937c4063c7dad562278ded49",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token unik perangkat timbangan",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "DEV-TOKEN-FG-PSN-001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses - status ready atau idle",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SettingsResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/fg-pasuruan/weight": {
            "post": {
                "tags": [
                    "FG Pasuruan"
                ],
                "summary": "Kirim data berat timbangan FG Pasuruan",
                "description": "Menerima data berat dari ESP8266/Arduino dan menyimpannya ke database. Broadcast via WebSocket ke dashboard real-time.",
                "operationId": "de62025726b03e40b5d4263e0268d3c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "token",
                                    "berat"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "Token unik perangkat",
                                        "type": "string",
                                        "example": "DEV-TOKEN-FG-PSN-001"
                                    },
                                    "berat": {
                                        "description": "Berat dalam kg",
                                        "type": "number",
                                        "format": "float",
                                        "example": 25.5
                                    },
                                    "kode_produksi": {
                                        "description": "Kode produksi (opsional jika sudah ada sesi aktif)",
                                        "type": "string",
                                        "example": "KP-2024-001"
                                    },
                                    "produk_id": {
                                        "description": "ID produk (opsional)",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berat berhasil disimpan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WeightSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Kode produksi belum diinput atau berat tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Tidak ada operator aktif",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/fg-pasuruan/device/update-product": {
            "post": {
                "tags": [
                    "FG Pasuruan"
                ],
                "summary": "Update produk yang sedang ditimbang oleh perangkat",
                "operationId": "2f425daab24f6ca36782089e5d94d1a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "token",
                                    "produk_id"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "DEV-TOKEN-FG-PSN-001"
                                    },
                                    "produk_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produk perangkat berhasil diupdate"
                    },
                    "401": {
                        "description": "Token tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SettingsResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "ready",
                            "idle"
                        ],
                        "example": "ready"
                    },
                    "kode_produksi": {
                        "type": "string",
                        "example": "KP-2024-001"
                    },
                    "nama_produk": {
                        "type": "string",
                        "example": "Mie Goreng 85g"
                    },
                    "operator": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "expired": {
                        "type": "string",
                        "example": "2025-12-31"
                    }
                },
                "type": "object"
            },
            "WeightSuccessResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Data berat 25.5 kg tersimpan"
                    },
                    "record_id": {
                        "type": "integer",
                        "example": 42
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            }
        }
    },
...
}